added remaining refactoring
