+++ app/__tests__/unit/features/NotificationWashingMachine.test.tsx
@@ -1,14 +1,14 @@
 import React from "react";
 import { render, fireEvent, act, waitFor, within } from "@testing-library/react-native";
-import WashingMachineScreen from "../../screens/laundry_machines/WashingMachineScreen";
+import WashingMachineScreen from "../../../screens/laundry_machines/WashingMachineScreen";
 import * as Notifications from "expo-notifications";
-import { getAllLaundryMachines, getLaundryMachinesQuery, updateLaundryMachine } from "../../../firebase/firestore/firestore";
+import { getAllLaundryMachines, getLaundryMachinesQuery, updateLaundryMachine } from "../../../../firebase/firestore/firestore";
 import { NotificationPermissionsStatus } from "expo-notifications";
 
 // portions of this code were generated by chatGPT as an AI assistant
 
 // Mock Firestore functions
-jest.mock("../../../firebase/firestore/firestore", () => ({
+jest.mock("../../../../firebase/firestore/firestore", () => ({
     getLaundryMachinesQuery: jest.fn(),
     updateLaundryMachine: jest.fn(),
     getLaundryMachine: jest.fn(() =>
